Remove idempotency key from UI flow
UI flow requires a restart in registration flow if the user entered validated but incorrect value (i.e. expiry date
or credit card number). Since after restarting we cannot reuse idempotency key, we cannot guarantee idempotency.
This is safe as double registration because of no idempotency key mechanism in ui is close to irrelevant as the flow will never be completed.